requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
pysparkpip3 